Extract SettingItem component
